ar pair tracking integrated in cafe controller new sound added
