+++ waiterbot_suite/waiterbot_ctrl_cafe/include/waiterbot_ctrl_cafe/ar_markers.hpp
@@ -19,6 +19,8 @@
 #include <geometry_msgs/PoseWithCovarianceStamped.h>
 #include <ar_track_alvar/AlvarMarkers.h>
 #include <yocs_ar_marker_tracking/tracking.hpp>
+#include <yocs_msgs/ARPair.h>
+#include <yocs_msgs/ARPairList.h>
 
 namespace waiterbot
 {
@@ -61,6 +63,27 @@ public:
 protected:
   void customCB(const ar_track_alvar::AlvarMarkers& spotted_markers, const std::vector<yocs::TrackedMarker> &tracked_markers);
 
+  void broadcastMarkersTF();
+    void publishMarkerTFs(const std::string prefix, const ar_track_alvar::AlvarMarkers& markers);
+    void publishTargetTFs(const std::string prefix, const ar_track_alvar::AlvarMarkers& markers);
+
+  void globalMarkersCB(const ar_track_alvar::AlvarMarkers::ConstPtr& msg);
+    void createFixedMarkers();
+    void createMirrorMarkers();
+    void notifyARPairTracker();
+
+
+  void arPoseMarkersCB(const ar_track_alvar::AlvarMarkers::ConstPtr& msg);
+
+  bool getMarkerTf(const std::string& ref_frame, const std::string& prefix, uint32_t marker_id,
+                     const ros::Time& timestamp, tf::StampedTransform& tf, const float timeout);
+  bool getTf(const std::string& ref_frame, const std::string& marker_frame, 
+               const ros::Time& timestamp, tf::StampedTransform& tf, const float timeout);
+
+  void print_transform(const std::string& name, tf::Transform& t);
+  void print_stampedtransform(const std::string& name, tf::StampedTransform& t);
+  void publish_transform(const std::string parent, const std::string child,const tf::Transform& t);
+
 private:
 
   // Confidence evaluation attributes
@@ -79,30 +102,23 @@ private:
   double                   tf_broadcast_freq_;  /**< Allows enabling tf broadcasting; mostly for debug */
 
   ar_track_alvar::AlvarMarker  docking_marker_;  /**< AR markers described in the semantic map */
-  ar_track_alvar::AlvarMarkers global_markers_,global_markers_mirrors;  /**< AR markers described in the semantic map */
+  ar_track_alvar::AlvarMarkers global_markers_, global_markers_fix_, global_markers_mirrors;  /**< AR markers described in the semantic map */
 
   bool               tracker_enabled_;
   ros::Subscriber    tracked_markers_sub_;
   ros::Subscriber    global_markers_sub_;
+  ros::Publisher     update_ar_pairs_pub_;
 
   // Registered callbacks to notify our observations
 
   boost::function<void (const geometry_msgs::PoseWithCovarianceStamped::ConstPtr&)>  robot_pose_cb_;
   boost::function<void (const geometry_msgs::PoseStamped::ConstPtr&, uint32_t)>    base_spotted_cb_;
 
-  // Private methods
-  void broadcastMarkersTF();
-  void globalMarkersCB(const ar_track_alvar::AlvarMarkers::ConstPtr& msg);
-  void arPoseMarkersCB(const ar_track_alvar::AlvarMarkers::ConstPtr& msg);
-
-  bool getMarkerTf(const std::string& ref_frame, const std::string& prefix, uint32_t marker_id,
-                     const ros::Time& timestamp, tf::StampedTransform& tf, const float timeout);
-  bool getTf(const std::string& ref_frame, const std::string& marker_frame, 
-               const ros::Time& timestamp, tf::StampedTransform& tf, const float timeout);
-
-  void print_transform(const std::string& name, tf::Transform& t);
-  void print_stampedtransform(const std::string& name, tf::StampedTransform& t);
-  void publish_transform(const std::string parent, const std::string child,const tf::Transform& t);
+  std::string global_marker_prefix_;
+  std::string target_frame_postfix_;
+  std::string global_marker_mirror_prefix_;
+  double target_offset_;
+  double baseline_;
 };
 
 } /* namespace waiterbot */
