+++ waiterbot_suite/waiterbot_ctrl_cafe/src/waiter_node.cpp
@@ -60,7 +60,7 @@ bool WaiterNode::init()
   }
 
   // Configure call-backs between submodules
-  ar_markers_.setRobotPoseCB(boost::bind(&NavWatchdog::arMarkerMsgCB, &nav_watchd_, _1));
+//  ar_markers_.setRobotPoseCB(boost::bind(&NavWatchdog::arMarkerMsgCB, &nav_watchd_, _1));
   ar_markers_.baseSpottedCB(boost::bind(&Navigator::baseSpottedMsgCB, &navigator_, _1, _2));
 
   // Enable safety controller by default
