+++ waiterbot_suite/waiterbot_ctrl_nowireless/launch/includes/_ar_tracker.launch
@@ -1,12 +1,12 @@
 <launch>
   <!-- openni is launched elswhere in the rapp (in rgb mode) -->
   <include file="$(find yocs_ar_marker_tracking)/launch/includes/ar_track_alvar.xml">
-    <arg name="ar_marker_size" value="5.4"/> <!-- Marcus': 5.4, Daniel's: 6.0 -->
+    <arg name="ar_marker_size" value="6.9"/> <!-- Marcus': 5.4, Daniel's: 6.0 -->
     <arg name="ar_marker_max_frequency" value="4.0"/>   <!-- This is a CPU-hungry bastard; reduce its rate  -->
   </include>
 
-  <arg name="ar_pair_left_id"            default="10"/>
-  <arg name="ar_pair_right_id"           default="11"/>
+  <arg name="ar_pair_left_id"            default="3"/>
+  <arg name="ar_pair_right_id"           default="0"/>
   <arg name="ar_pair_baseline"/> <!-- YO!!!! Change this in tandem with the next one! -->
   <arg name="ar_pair_baseline_half"/>
   <arg name="ar_pair_target_pose_offset"/> <!-- Target distance in front of the midpoint of the two ar markers -->
